feat: enviar cada 2 minutos un email de  tipo primera promocion
